Delete to_js_lang_name from coursetags.py
