+++ course/templatetags/coursetags.py
@@ -24,21 +24,11 @@ THE SOFTWARE.
 
 from django.template import Library, Node, TemplateSyntaxError
 from django.utils import translation
+from relate.utils import to_js_lang_name
 
 register = Library()
 
 # {{{ get language_code in JS traditional naming format
-
-def to_js_lang_name(dj_lang_name):
-    """
-    Turns a django language name (en-us) into a js styled language
-    name (en-US).
-    """
-    p = dj_lang_name.find('-')
-    if p >= 0:
-        return dj_lang_name[:p].lower() + '-' + dj_lang_name[p + 1:].upper()
-    else:
-        return dj_lang_name.lower()
 
 class GetCurrentLanguageJsFmtNode(Node):
     def __init__(self, variable):
